# Conflicts:
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/jy/dao/pickinggood/JyBizTaskPickingGoodDao.java
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/jy/service/picking/JyBizTaskPickingGoodService.java
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/jy/service/picking/JyBizTaskPickingGoodServiceImpl.java

+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/jy/service/picking/JyPickingSendDestinationService.java
@@ -57,5 +57,5 @@ public interface JyPickingSendDestinationService {
      */
     boolean deleteSendFlow(SendFlowDeleteReq req);
 
-    List<JyPickingSendDestinationDetailEntity> listSendCodesByCreateSiteId(Long createSiteId, List<Long> destinationSiteId);
+    List<JyPickingSendDestinationDetailEntity> listSendCodesByCreateSiteId(Long createSiteId, List<Integer> destinationSiteId);
 }
